Quick hack for Terraria 1.4.1.1
This is a quick hack for macOS and Terraria 1.4.1.1.

The biggest change is that the world version number has increased from
230 to 233, and changing this lets the game load.

Also, it appears that some of the files have moved, at least on macOS,
so I tweaked the default locations in the settingsdialog.cpp. You can
manually override the settings, but IMHO that should not be necessary
for an unmodified Steam install.

TODO:
- Make sure any new textures are supported
- Possibly add a check in settingsdialog.cpp to look for files in both
  the new and old places, and if it only exists in one place, use that.
- See what changes, if any, are needed for Windows and Linux